# api/openapi.yaml
openapi: 3.0.3
info:
  title: recommend4me
  version: 0.1.0
  description: |
    A local recommender of works — games, books — collected from sites by plugins. The user grades
    some of them; a model learns the order of the grades and puts every other work on the user's
    own 0..10 scale.

    An error is answered as application/problem+json (the Problem schema): its "detail" says, in
    the user's language, what went wrong.
...
paths:
...
  /api/types/{type}/items:
    get:
      tags: [items]
      operationId: listItems
      summary: A page of the works of a content type, one card per work
      parameters:
        - $ref: '#/components/parameters/Type'
        - name: view
          in: query
          schema:
            $ref: '#/components/schemas/ItemView'
        - name: sort
          in: query
          schema:
            $ref: '#/components/schemas/ItemSort'
        - name: sortNumber
          in: query
          description: The number to sort by when sort is NUMBER (a facet id of the numbers)
          schema:
            type: string
        - name: search
          in: query
          description: A line searched by meaning and by words; the order is then the relevance
          schema:
            type: string
        - name: hidden
          in: query
          description: Values of filter facets not shown, as "<facet id>=<value key>"
          schema:
            type: array
            items:
              type: string
        - name: hiddenWithout
          in: query
          description: Filter facets whose works without any value of them are not shown
          schema:
            type: array
            items:
              type: string
        - name: source
          in: query
          description: Only the works of these sources; all when none
          schema:
            type: array
            items:
              type: string
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 60
      responses:
        '200':
          description: The page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemPage'
  /api/types/{type}/facets:
    get:
      tags: [items]
      operationId: listFacets
      summary: The filter facets of a content type with their values and how many works have each
      parameters:
        - $ref: '#/components/parameters/Type'
      responses:
        '200':
          description: The facets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FacetFilter'
  /api/types/{type}/lookup:
    get:
      tags: [items]
      operationId: lookupItems
      summary: Works of a content type by a piece of the title or by the address of the work's page, each item apart
      description: |
        Unlike the list, every item is found by itself, linked to others or not: what the user
        looks for when saying two items are one work.
      parameters:
        - $ref: '#/components/parameters/Type'
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: The items found, the address first, then by the title
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LinkedItem'
  /api/sources/{source}/facets/{facet}/values:
    get:
      tags: [items]
      operationId: listFacetValues
      summary: The values of a facet of a source, by how many items have each, for a value to be added
      parameters:
        - $ref: '#/components/parameters/Source'
        - name: facet
          in: path
          required: true
          schema:
            type: string
        - name: query
          in: query
          description: A piece of the name, any case; all when none
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: The values, the most used first; the ones the user made up among them
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FacetValueUse'
        '404':
          description: No such source or facet
  /api/items/{source}/{item}:
    get:
      tags: [items]
      operationId: getItem
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
      responses:
        '200':
          description: Everything known of the item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetails'
        '404':
          description: No such item
...
  /api/items/{source}/{item}/refresh:
    post:
      tags: [items]
      operationId: refreshItem
      summary: Downloads the item's page again now, if its source can
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
      responses:
        '200':
          description: The item after the download
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetails'
        '404':
          description: No such item
        '409':
          description: The source cannot download pages
        '502':
          description: The site did not answer
  /api/items/{source}/{item}/rating:
    put:
      tags: [items]
      operationId: rateItem
      summary: Grades the item's current version
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingRequest'
      responses:
        '200':
          description: The item's card after the grade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemSummary'
    delete:
      tags: [items]
      operationId: unrateItem
      summary: Takes back the grade of the item's current version, or of the version named
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
        - name: version
          in: query
          description: An earlier version whose grade is taken back; the current one when none
          schema:
            type: string
      responses:
        '200':
          description: The item's card without the grade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemSummary'
  /api/items/{source}/{item}/pictures/{position}:
    get:
      tags: [items]
      operationId: getPicture
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
        - $ref: '#/components/parameters/Position'
        - name: full
          in: query
          description: The original instead of the preview
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: The picture
          content:
            image/*:
              schema:
                type: string
                format: binary
        '302':
          description: Not kept here; the site's address
        '404':
          description: No such picture
  /api/items/{source}/{item}/pictures/{position}/mark:
    put:
      tags: [items]
      operationId: markPicture
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
        - $ref: '#/components/parameters/Position'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mark'
      responses:
        '204':
          description: Marked
        '404':
          description: No such picture
    delete:
      tags: [items]
      operationId: unmarkPicture
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
        - $ref: '#/components/parameters/Position'
      responses:
        '204':
          description: Unmarked
  /api/items/{source}/{item}/reviews/{reviewId}/mark:
    put:
      tags: [items]
      operationId: markReview
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
        - $ref: '#/components/parameters/ReviewId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mark'
      responses:
        '204':
          description: Marked
        '404':
          description: No such review
    delete:
      tags: [items]
      operationId: unmarkReview
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
        - $ref: '#/components/parameters/ReviewId'
      responses:
        '204':
          description: Unmarked
  /api/items/{source}/{item}/parts/{partId}:
    get:
      tags: [items]
      operationId: getPart
      summary: The text of a part of the work, as it was read
      parameters:
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Item'
        - name: partId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The part
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartText'
        '404':
          description: No such part, or its text is not kept
  /api/match-feedback:
    post:
      tags: [items]
      operationId: giveMatchFeedback
      summary: The user's word on a match shown for a mark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchFeedbackRequest'
      responses:
        '204':
          description: Kept
...
components:
  parameters:
    Type:
      name: type
      in: path
      required: true
      schema:
        type: string
    Source:
      name: source
      in: path
      required: true
      schema:
        type: string
    Item:
      name: item
      in: path
      required: true
      schema:
        type: string
    Position:
      name: position
      in: path
      required: true
      schema:
        type: integer
    ReviewId:
      name: reviewId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Grade:
      type: integer
      minimum: 1
      maximum: 5
      description: |
        The user's grade, an order and not a number: 1 - do not like it, 2 - can be played (read),
        3 - liked it on the whole, 4 - very good, 5 - give me more.
...
    ItemView:
      type: string
      description: UNRATED - works whose current version has no grade; RATED - graded; ALL - everything
      enum: [UNRATED, RATED, ALL]
      default: UNRATED
    ItemSort:
      type: string
      description: SCORE - by prediction; UPDATED - newest update first; NUMBER - by a number, the largest first
      enum: [SCORE, UPDATED, NUMBER]
      default: SCORE
    FacetValueInfo:
      type: object
      required: [key, name]
      properties:
        key:
          type: string
        name:
          type: string
        corrected:
          type: string
          enum: [ADDED, REMOVED]
          description: The user's correction of the site's value; a REMOVED value is shown crossed out
    ItemFacet:
      type: object
      required: [facet, label, values]
      properties:
        facet:
          type: string
          description: The facet's key within its source
        label:
          type: string
        values:
          type: array
          items:
            $ref: '#/components/schemas/FacetValueInfo'
    Prediction:
      type: object
      required: [score]
      properties:
        score:
          type: number
          format: double
          description: |
            The work's place on the user's own scale: the centre of the lowest grade is 0, of the
            highest 10; works beyond the extreme grades go beyond 0..10
...
    LinkedItem:
      type: object
      required: [source, item, title, url]
      properties:
        source:
          type: string
        item:
          type: string
        title:
          type: string
        url:
          type: string
    ItemSummary:
      type: object
      required: [source, item, title, version, url, updatedAt, facets, numbers, hasCover, pictureCount, linked, work]
      properties:
        work:
          type: string
          description: |
            The work the item is of, as "<source>/<item>" of the item that stands for it in the
            list; the item's own key when it is linked to no other
        source:
          type: string
        item:
          type: string
        title:
          type: string
        version:
          type: string
        url:
          type: string
        updatedAt:
          type: string
          format: date-time
        facets:
          type: array
          description: The facets shown on a card
          items:
            $ref: '#/components/schemas/ItemFacet'
        numbers:
          type: object
          description: By the number's key within the item's source (SourceInfo.numbers[].key), corrections applied
          additionalProperties:
            type: number
            format: double
        hasCover:
          type: boolean
        pictureCount:
          type: integer
        grade:
          description: The grade of the work's current version - given to this item or to another item of the same work
          allOf:
            - $ref: '#/components/schemas/Grade'
        previousGrade:
          $ref: '#/components/schemas/Grade'
        previousGradeVersion:
          type: string
        prediction:
          $ref: '#/components/schemas/Prediction'
        searchMatch:
          $ref: '#/components/schemas/SearchMatch'
        signals:
          type: object
          description: The user's own actions on the site (liked, on a shelf)
          additionalProperties:
            type: string
        linked:
          type: array
          description: The other items that are this same work
          items:
            $ref: '#/components/schemas/LinkedItem'
    SearchMatch:
      type: object
      required: [field, byMeaning, text, highlights]
      properties:
        field:
          type: string
          description: The name of the field it was found in, as the user is shown it
        byMeaning:
          type: boolean
          description: Found by meaning alone; the text is then the phrase of the field nearest by meaning
        text:
          type: string
        highlights:
          type: array
          items:
            $ref: '#/components/schemas/TextRange'
...
    ItemPage:
      type: object
      required: [total, items]
      properties:
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemSummary'
    FacetFilter:
      type: object
      required: [id, label, itemsWithout, values]
      properties:
        id:
          type: string
        label:
          type: string
        noneLabel:
          type: string
          description: The name of the works without a value of the facet; none - they are not a line of the filter
        itemsWithout:
          type: integer
        values:
          type: array
          items:
            $ref: '#/components/schemas/FacetValueUse'
    FacetValueUse:
      type: object
      required: [key, name, items]
      properties:
        key:
          type: string
        name:
          type: string
        items:
          type: integer
    TextValue:
      type: object
      required: [key, label, content, spoiler, corrected]
      properties:
        key:
          type: string
        label:
          type: string
        content:
          type: string
        spoiler:
          type: boolean
        corrected:
          type: boolean
    NumberValue:
      type: object
      required: [key, label, value, corrected]
      properties:
        key:
          type: string
        label:
          type: string
        value:
          type: number
          format: double
        corrected:
          type: boolean
    PartInfo:
      type: object
      required: [partId, position, hasText]
      properties:
        partId:
          type: string
        position:
          type: integer
        title:
          type: string
        hasText:
          type: boolean
        publishedAt:
          type: string
          format: date-time
    PartText:
      type: object
      required: [partId, position, content]
      properties:
        partId:
          type: string
        position:
          type: integer
        title:
          type: string
        content:
          type: string
    ItemDetails:
      type: object
      required: [summary, allFacets, texts, allNumbers, explanation, ratings, pictures, reviews, parts, pictureMatches, reviewMatches, titleCorrected, canRefresh]
      properties:
        summary:
          $ref: '#/components/schemas/ItemSummary'
        titleCorrected:
          type: boolean
        allFacets:
          type: array
          description: Every facet of the item, corrections included
          items:
            $ref: '#/components/schemas/ItemFacet'
        texts:
          type: array
          items:
            $ref: '#/components/schemas/TextValue'
        allNumbers:
          type: array
          items:
            $ref: '#/components/schemas/NumberValue'
        explanation:
          type: array
          description: Features that pushed the prediction most, strongest first
          items:
            $ref: '#/components/schemas/FeatureContribution'
        ratings:
          type: array
          description: The grades of every item of the work, newest first
          items:
            $ref: '#/components/schemas/RatingRecord'
        reviews:
          type: array
          description: The reviews that move the work most either way, strongest first; or the newest
          items:
            $ref: '#/components/schemas/ReviewInfo'
        reviewCount:
          type: integer
        pictures:
          type: array
          items:
            $ref: '#/components/schemas/PictureInfo'
        parts:
          type: array
          items:
            $ref: '#/components/schemas/PartInfo'
        pictureMatches:
          type: array
          items:
            $ref: '#/components/schemas/PictureMatch'
        reviewMatches:
          type: array
          items:
            $ref: '#/components/schemas/ReviewMatch'
        canRefresh:
          type: boolean
          description: The source can download the item's page again
    MatchFeedbackRequest:
      type: object
      required: [kind, markSource, markItem, markRef, matchSource, matchItem, matchRef, verdict]
      properties:
        kind:
          type: string
          enum: [PICTURE, REVIEW]
        markSource:
          type: string
        markItem:
          type: string
        markRef:
          type: string
          description: The marked picture's position or the marked review's id
        matchSource:
          type: string
        matchItem:
          type: string
        matchRef:
          type: string
        verdict:
          type: integer
          minimum: -1
          maximum: 1
          description: +1 alike in the sense meant, -1 alike in another sense, 0 takes the word back
    PictureMatch:
      type: object
      required: [position, markSource, markItem, markTitle, markPosition, mark, strength]
      properties:
        position:
          type: integer
        markSource:
          type: string
        markItem:
          type: string
        markTitle:
          type: string
        markPosition:
          type: integer
        verdict:
          type: integer
        mark:
          type: integer
        strength:
          type: number
          format: double
          description: |
            How surprising the likeness is: -ln of the chance that as many random pictures of the
            catalogue would come as close; chance gives about 1, 3 one work in twenty, 5 one in 150
    ReviewMatch:
      type: object
      required: [reviewId, content, markSource, markItem, markReviewId, markTitle, markContent, mark, strength]
      properties:
        reviewId:
          type: string
        content:
          type: string
        markSource:
          type: string
        markItem:
          type: string
        markReviewId:
          type: string
        markTitle:
          type: string
        markContent:
          type: string
        verdict:
          type: integer
        mark:
          type: integer
        strength:
          type: number
          format: double
    ReviewInfo:
      type: object
      required: [reviewId, content]
      properties:
        reviewId:
          type: string
        author:
          type: string
        stars:
          type: number
          format: double
        content:
          type: string
        postedAt:
          type: string
          format: date-time
        influence:
          type: number
          format: double
          description: Points of the 0..10 scale the review moves the work, left out of the reviews' set and of their likeness to the marked ones
        mark:
          type: integer
    Mark:
      type: object
      required: [mark]
      properties:
        mark:
          type: integer
          minimum: -1
          maximum: 1
          description: +1 a reason to pick the work, -1 a reason to drop it
    PictureInfo:
      type: object
      required: [position, hasFull, analyzed]
      properties:
        position:
          type: integer
          description: 0 the cover, 1.. the screenshots
        hasFull:
          type: boolean
        analyzed:
          type: boolean
        mark:
          type: integer
        influence:
          type: number
          format: double
        error:
          type: string
    RatingRecord:
      type: object
      required: [source, item, version, grade, ratedAt, current]
      properties:
        source:
          type: string
        item:
          type: string
          description: The item of the work graded - this one or one linked to it
        current:
          type: boolean
          description: The grade is of the item's current version
        version:
          type: string
        grade:
          $ref: '#/components/schemas/Grade'
        ratedAt:
          type: string
          format: date-time
    FeatureContribution:
      type: object
      required: [feature, label, contribution, present]
      properties:
        feature:
          type: string
        label:
          type: string
        contribution:
          type: number
          format: double
          description: Points of the 0..10 scale the feature moves the work
        present:
          type: boolean
    RatingRequest:
      type: object
      required: [grade]
      properties:
        grade:
          $ref: '#/components/schemas/Grade'
...